requirements.txt
# Vulnerable dependencies for Trivy to detect
requests==2.19.1  # CVE-2018-18074 - Information disclosure
urllib3==1.24.1   # Multiple CVEs
pyyaml==5.3.1     # CVE-2020-14343 - Arbitrary code execution
flask==0.12.2     # CVE-2018-1000656 - Denial of service
